fix(build): revert most error prone link switch
our ci runs through, but some errors still appeared for @aschaeffer (grpc: gpr linking errors in inexor-core-client, although it was in grpcs libs list)
the quickfix is: revert the change in "rename binaries to.." which removed a line which adds all conan libs to any binary.
we will have to solve this after we modularized more.
the quickfix only increases the time needed to link, no other harm made.